+++ package_release.ps1
@@ -1,4 +1,4 @@
-$version = "1.1.9"
+$version = "1.2.0"
 $appName = "SocialRadar"
 $buildDirName = "chrome-mv3-prod"
 $targetName = "$appName-v$version"